Addresses https://github.com/moneypenny/gulp-ruby-haml/issues/10
This patch will allow the haml() plugin to process haml without requiring it to exist on the physical disk.

Instead of supplying the haml command with the path of an input file, we supply the switch -s, and write the haml to childprocess.stdin. This lets us do the following

<pre><code>
gulp.src('foo/bar/**.*.haml')
  .pipe(replace('albert', 'dilbert')
  .pipe(haml())
  .pipe(gulp.dest('baz'));
</code></pre>